Move genes to gene.py
