+++ roxene/genes.py
@@ -1,6 +1,6 @@
 from tensorflow import Tensor
 
-from roxene import Gene, Organism, Neuron
+from roxene import Gene, Organism, Neuron, Cell
 
 
 class CreateNeuron(Gene):
@@ -33,3 +33,27 @@ class CreateNeuron(Gene):
             hidden_output=self.hidden_output
         )
         organism.add(neuron)
+
+
+class ConnectNeurons(Gene):
+
+    def __init__(self, tx_cell_index, rx_input_port):
+        self.tx_cell_index = tx_cell_index
+        self.rx_port = rx_input_port
+
+
+    def execute(self, organism: Organism):
+        index: int = self.tx_cell_index % len(organism.cells)
+        tx_cell: Cell = organism.cells[index]
+        rx_cell: Neuron = organism.cells[0]
+        rx_cell.add_input_connection(tx_cell, self.rx_port)
+
+
+class RotateCells(Gene):
+
+    def __init__(self, direction):
+        self.direction = direction
+        pass
+
+    def execute(self, organism: Organism):
+        organism.cells.rotate(self.direction)
